requirements.txt
openai
pinecone-client
tiktoken
langchain
python-dotenv
PyJWT
fleet-context>=1.1.7